feat: add save heart and toggle functionality - separate random_words functionality into separate file - add heart icon beside each random word - add functionality to toggle if random word is saved
