+++ wordpair_generator/lib/random_words.dart
@@ -1,0 +1,63 @@
+import 'package:flutter/material.dart';
+import 'package:english_words/english_words.dart';
+
+class RandomWords extends StatefulWidget {
+  @override
+  RandomWordsState createState() => RandomWordsState();
+}
+
+class RandomWordsState extends State<RandomWords> {
+  final _randomWordPairs = <WordPair>[]; // defined as a list
+  final _savedWordPairs = Set<WordPair>();
+
+  Widget _buildList() {
+    // If using itemBuilder, include `.builder` after ListView
+    return ListView.builder( 
+      padding: const EdgeInsets.all(16.0),
+      itemBuilder: (context, item) {
+        if(item.isOdd) return Divider();
+        
+        final index = item ~/ 2; // ~/ divide an integer
+
+        if (index >= _randomWordPairs.length) {
+          // Use generateWordPairs from english_words
+          _randomWordPairs.addAll(generateWordPairs().take(10)); 
+        }
+
+        return _buildRow(_randomWordPairs[index]); 
+      }
+    );
+  }
+
+  Widget _buildRow(WordPair pair) {
+    final alreadySaved = _savedWordPairs.contains(pair); 
+    return ListTile(
+      title: Text(
+        pair.asPascalCase, 
+        style: TextStyle(fontSize: 18.0)
+      ),
+      trailing: Icon(
+        alreadySaved ? Icons.favorite : Icons.favorite_border, 
+        color: alreadySaved ? Colors.red : null
+      ),
+      onTap: () {
+        setState(() {
+          if(alreadySaved) {
+            _savedWordPairs.remove(pair);
+          } else {
+            _savedWordPairs.add(pair);
+          }
+        });
+      }
+    );
+  }
+
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: _buildList(),
+      appBar: AppBar(
+        title: Text('Wordpair Generator')
+      )
+    );
+  }
+}
